Changed file/folder structure, Changed from big class to namespaces
