+++ src/ElementProperties/touch.h
@@ -1,0 +1,21 @@
+#pragma once
+
+namespace drawIt_intern {
+
+class touch {
+   public:
+    virtual bool touched(uint16_t x, uint16_t y) = 0;
+
+    void setTouch(bool activated) {
+        _touch = activated;
+    }
+
+    bool getTouch() {
+        return _touch;
+    }
+
+   protected:
+    bool _touch;
+};
+
+}  // namespace drawIt_intern
